+++ project/frontend/src/components/Pages/SignupPage.js
@@ -1,5 +1,6 @@
 import React from "react";
+import SignUp from "../Authentication/Signup/Signup";
 
-const SignupPage = () => (<></>)
+const SignupPage = () => (<><SignUp /></>)
 
 export default SignupPage;
