+++ project/frontend/src/components/Pages/SignupPage.test.js
@@ -1,6 +1,7 @@
 import React from "react";
 import SignupPage from "./SignupPage";
 import { shallow } from "enzyme";
+import Signup from "../Authentication/Signup/Signup";
 
 describe("<SignupPage />", () => {
   it("should contain <Signup /> component", () => {
